+++ index.js
@@ -1,7 +1,7 @@
 global.__TEST__ = false
 
 import { AppRegistry } from "react-native"
-import { App } from "./src/App"
+import { App } from "./app/App"
 import { name as appName } from "./app.json"
 
 AppRegistry.registerComponent(appName, () => App)
